// src/enums/ScriptsEnum.ts
export enum ScriptsEnum {
  flashcardsQuestions = "flashcards/questions",
  flashcardsResponses = "flashcards/responses",
  quizzesFeedbacks = "quizzes/feedbacks",
  quizzesQuestions = "quizzes/questions",
  all = "all",
  updateStories = "update/stories",
  clearAudio = "clear/audio",
  clearQuizQuestions = "clear/quizzes/questions" // Nouvelle entrée
}